measuring: add some code to measure performance
Comparing two cases: one where the contents of every file is read and one
where we skip the read but do everything else. Results:

         Files      Bytes  Time(secs)     Mb/sec         files/sec
READ    373068 34719401787 128.848321  269.459481641  2895.40443449
NO READ 373067 34719394657   5.198403 6678.8578448   71765.6941949

So there is a strong case to avoid reading if possible.  Proposed solution
is to use data from the previous back up and it st_mtime and st_size haven't
changed then assume that they haven't changed and use the previous digest.

Signed-off-by: Peter Williams <pwil3058@gmail.com>