gzip: measure the cost/benefit of gzipping the snapshot file
Cost: adds 64.090947 seconds to load and 20.753151 seconds to dump time
Benefit: reduces size of snapshot from 115102691 to 24993622 (0.21)

Means command line queries would be very slow as the snapshot file would
have to be read for each query.  However, a GUI can just read it once and
answer any questions with a cached copy as it isn't going to change on disk.
There would be a delay in changing snapshots being queried.

Signed-off-by: Peter Williams <pwil3058@gmail.com>